break dependency of filaflat on filabridge
filaflat only had on header dependency on filabridge (DriverEnums.h)
and only needed two small enum types.
In fact, I don't think it was right for filaflat to assume any 
particular value for these fields -- this is the responsibility of the
callers.

+++ tools/matinfo/src/main.cpp
@@ -113,13 +113,13 @@ public:
         switch (mBackend) {
             case Backend::OPENGL:
                 return materialChunk.getTextShader(unflattener, blobDictionary,
-                        shader, shaderModel, variant, st);
+                        shader, (uint8_t)shaderModel, variant, st);
             case Backend::METAL:
                 return materialChunk.getTextShader(unflattener, blobDictionary,
-                        shader, shaderModel, variant, st);
+                        shader, (uint8_t)shaderModel, variant, st);
             case Backend::VULKAN:
                 return materialChunk.getSpirvShader(unflattener, blobDictionary,
-                        shader, shaderModel, variant, st);
+                        shader, (uint8_t)shaderModel, variant, st);
             default:
                 return false;
         }
